requirements.txt
Django==2.0
djangorestframework
markdown
django-filter
django-extensions